+++ packages/app/src/app/account/[id]/predicate/page.tsx
@@ -1,5 +1,5 @@
 import { getPredicate } from '~/systems/Account/actions/get-predicate';
-import { AccountPredicate } from '~/systems/Account/screens/AccountPredicate/AccountPredicate';
+import { AccountPredicate } from '~/systems/Account/screens/AccountPredicate';
 
 type PageProps = {
   params: {
@@ -11,10 +11,7 @@ export default async function AccountPredicatePage({
   params: { id },
 }: PageProps) {
   const predicate = await getPredicate({ owner: id });
-
-  if (!predicate?.bytecode) return null;
-
-  return <AccountPredicate bytecode={predicate.bytecode} />;
+  return <AccountPredicate bytecode={predicate?.bytecode ?? ''} />;
 }
 
 export const revalidate = 100;
