--- packages/app/src/systems/Account/screens/AccountScreen/AccountScreen.tsx
@@ -1,22 +1,0 @@
-'use client';
-
-import type { AccountBalanceFragment, Maybe } from '@fuel-explorer/graphql';
-import { VStack } from '@fuels/ui';
-
-import { BalanceItem } from '../../../Core/components/BalanceItem/BalanceItem';
-
-type AccountScreenProps = {
-  balances?: Maybe<AccountBalanceFragment[]>;
-};
-
-export function AccountScreen({ balances }: AccountScreenProps) {
-  return (
-    <VStack className="min-h-[45vh]">
-      {balances?.map((balance) => {
-        return (
-          <BalanceItem key={balance.assetId + balance.owner} item={balance} />
-        );
-      })}
-    </VStack>
-  );
-}
